+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - feat(api_security_operations): add `fastly_api_security_operations` and `fastly_api_security_operation_tags` data sources ([#101](https://github.com/fastly/terraform-provider-fastly-beta/pull/101))
 - feat(package_hash): add `fastly_package_hash` data source ([#103](https://github.com/fastly/terraform-provider-fastly-beta/pull/103))
 - feat(services): add `fastly_services` data source ([#102](https://github.com/fastly/terraform-provider-fastly-beta/pull/102))
+- feat(elasticsearch_logging): add support for Elasticsearch Logging ([#105](https://github.com/fastly/terraform-provider-fastly-beta/pull/105))
 
 ### BUG FIXES: